// src/Interfaces/appInterfaces.tsx
// Generated by https://quicktype.io

export interface LoginResponse {
  usuario: Usuario;
  token: string;
}
export interface RegisterData {
  correo: string;
  password: string;
  nombre: string;
}
export interface Usuario {
  rol: string;
  estado: boolean;
  google: boolean;
  nombre: string;
  correo: string;
  uid: string;
  img?: string;
}
export interface LoginData {
  correo: string;
  password: string;
}

export interface CursosDisponiblesResponse {
  label: string;
  value: string;
}
export interface Location {
  latitude: number;
  longitude: number;


}

export interface MetodosPago {
  yape: string;
  plin: string;
  efectivo: string;
}

// Generated by https://quicktype.io

export interface ProfesorCercano {
  nombre: string;
  tarifa: string;
  calificacion: string;
  tiempo: string;
  totalCalificaciones: string;
  id: string;
  pantalla: string
}


// Generated by https://quicktype.io

// Generated by https://quicktype.io

// Generated by https://quicktype.io

export interface OpinionesUsuarios {
  totalOpiniones: string;
  opiniones: Opinion[];
}

export interface Opinion {
  nombre: string;
  calificacion: string;
  id: string;
  opinion: string;
  reacciones: string;
  titulo: string;
}


// Generated by https://quicktype.io

export interface PagosRecibidosEnviadosResponse {
  id: string;
  pagos_enviados: PagosEnviados[];
  pagos_recibidos: PagosRecibidos[];
}

export interface PagosEnviados {
  para: string;
  monto: string;
  id_profesor: string;
  fecha: string;
}

export interface PagosRecibidos {
  de: string;
  monto: string;
  id_alumno: string;
  fecha: string;
}


// Generated by https://quicktype.io

export interface Courses {
  name: string;
  id: string;
  icon: string;
  background: string
}